requirements.txt
pyrogram==2.0.35
tgcrypto
pymongo[srv]==3.12.3
motor==2.5.1
marshmallow==3.14.1
umongo==3.0.1
requests
bs4
imdbpy==2022.7.9
aiohttp
pytz
hachoir
Pillow
humanize
aiofiles
dnspython
pynewtonmath
openai